+++ src/main/java/com/planner/planner/RowMapper/PlanLocationRowMapper.java
@@ -16,8 +16,8 @@ public class PlanLocationRowMapper implements RowMapper<PlanLocationDto> {
 				.setLocationContentId(rs.getInt("location_content_id"))
 				.setLocationImage(rs.getString("location_image"))
 				.setLocationAddr(rs.getString("location_addr"))
-				.setLocationMapx(rs.getFloat("location_mapx"))
-				.setLocationMapy(rs.getFloat("location_mapy"))
+				.setLocationMapx(rs.getDouble("location_mapx"))
+				.setLocationMapy(rs.getDouble("location_mapy"))
 				.setLocationTransportation(rs.getInt("location_transportation"))
 				.setPlanId(rs.getInt("plan_id"))
 				.build();
